bugs fixed

+++ mytorch/activation/relu.py
@@ -4,14 +4,12 @@ from mytorch import Tensor, Dependency
 def relu(x: Tensor) -> Tensor:
     "TODO: implement relu function"
 
-    # use np.maximum
     data = np.maximum(0, x.data)
     req_grad = x.requires_grad
 
     if req_grad:
         def grad_fn(grad: np.ndarray):
-            # use np.where
-            return grad * (x.data > 0).astype(np.float64)
+            return grad * (x.data > 0)
 
         depends_on = [Dependency(x, grad_fn)]
     else:
